# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hachiko API",
    "version": "0.1.0",
    "description": "API del e-commerce Hachiko. Autenticación por cookies httpOnly (JWT firmado HS256): `hachiko_access` (15 min) y `hachiko_refresh` (7 días, scope `/api/auth/refresh`). Todas las respuestas de error siguen el formato `{ ok: false, error: { code, message } }`.",
    "contact": {
      "email": "contacto@hachiko.cl"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    },
    {
      "url": "https://hachiko.cl",
      "description": "Producción"
    }
  ],
...
  "paths": {
...
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Solicitar recuperación de contraseña",
        "description": "Envía por correo un enlace de un solo uso (válido 60 min) para crear una contraseña nueva. La respuesta es idéntica exista o no la cuenta (anti-enumeración). Rate limit: 5 por IP cada 15 minutos.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitud aceptada (misma respuesta exista o no el email)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOnly"
                }
              }
            }
          },
          "400": {
            "description": "Email inválido (code: VALIDATION)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Demasiados intentos (code: RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno (code: INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Restablecer la contraseña con un token",
        "description": "Consume el token recibido por correo (un solo uso, atómico) y fija la contraseña nueva. Revoca todas las sesiones (tokenVersion) y desbloquea la cuenta. Rate limit: 10 por IP por minuto.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "password"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 32,
                    "maxLength": 128
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "description": "Misma política del registro: mayúscula y número"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contraseña actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOnly"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o token expirado/usado (codes: VALIDATION, INVALID_TOKEN)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Demasiados intentos (code: RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno (code: INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hachiko_access",
        "description": "JWT de acceso (15 min) en cookie httpOnly, SameSite=Strict"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hachiko_refresh",
        "description": "JWT de refresco (7 días) con path restringido a /api/auth/refresh"
      },
      "cronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET — solo lo envía el scheduler de Vercel"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Código estable para manejo programático"
              },
              "message": {
                "type": "string",
                "description": "Mensaje legible en español"
              }
            }
          }
        }
      },
      "OkOnly": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          }
        }
      },
...
    }
  }
}